pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "e220_900t22s"
version = "1.0.0"
authors = [
  { name="Taogya", email="132679709+taogya@users.noreply.github.com" },
]
description = "E220-900T22S Python Library for Raspberry Pi."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pyserial >= 3.5",
    "RPi-GPIO-Helper >= 2.0.0",
]

[tool.hatch.build]
exclude = [
    "build",
]

[project.urls]
"Homepage" = "https://github.com/taogya/RPi_E220-900T22S_LoRa_library"
"Bug Tracker" = "https://github.com/taogya/RPi_E220-900T22S_LoRa_library/issues"
